# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "FST Platform 后端 API 接口文档",
        "title": "FST Platform API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/v1/admin/settings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取所有系统配置，按分类分组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-设置"
                ],
                "summary": "获取所有系统配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/app_controllers_admin_settings.SettingsListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建一个新的自定义配置项",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-设置"
                ],
                "summary": "创建新配置",
                "parameters": [
                    {
                        "description": "配置信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app_controllers_admin_settings.CreateSettingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/v1/admin/settings/batch": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "批量更新多个配置项的值",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-设置"
                ],
                "summary": "批量更新配置",
                "parameters": [
                    {
                        "description": "配置键值对",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app_controllers_admin_settings.BatchUpdateSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/v1/admin/settings/category/{category}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定分类下的所有配置项",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-设置"
                ],
                "summary": "获取指定分类的配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分类名称",
                        "name": "category",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/v1/admin/settings/restart-backend": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Admin-设置"
                ],
                "summary": "重启后端进程",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/v1/admin/settings/server-monitoring": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Admin-设置"
                ],
                "summary": "服务端运行监控",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/v1/admin/settings/server-ops": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Admin-设置"
                ],
                "summary": "服务运营状态",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/v1/admin/settings/{key}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据键名获取配置详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-设置"
                ],
                "summary": "获取单个配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "配置键名",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新指定配置项的值",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-设置"
                ],
                "summary": "更新单个配置值",
                "parameters": [
                    {
                        "type": "string",
                        "description": "配置键名",
                        "name": "key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "配置值",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app_controllers_admin_settings.UpdateSettingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除指定的自定义配置项",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-设置"
                ],
                "summary": "删除配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "配置键名",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/v1/admin/settings/{key}/meta": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新配置项的完整信息（包括值和元数据）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-设置"
                ],
                "summary": "更新配置元数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "配置键名",
                        "name": "key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "配置信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app_controllers_admin_settings.UpdateSettingMetaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/v1/user/settings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前用户的个人设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User-设置"
                ],
                "summary": "获取用户设置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新当前用户的个人设置（语言、主题、通知偏好等）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User-设置"
                ],
                "summary": "更新用户设置",
                "parameters": [
                    {
                        "description": "设置信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app_controllers_user_profile.UpdateSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app_controllers_admin_settings.BatchUpdateSettingsRequest": {
            "type": "object",
            "required": [
                "settings"
            ],
            "properties": {
                "settings": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "app_controllers_admin_settings.CreateSettingRequest": {
            "type": "object",
            "required": [
                "key",
                "label"
            ],
            "properties": {
                "category": {
                    "description": "basic, security, email, custom",
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "is_editable": {
                    "type": "boolean"
                },
                "is_public": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                },
                "type": {
                    "description": "string, number, boolean, json",
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "app_controllers_admin_settings.SettingsListResponse": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fst_backend_app_models.SettingsGroup"
                    }
                }
            }
        },
        "app_controllers_admin_settings.UpdateSettingMetaRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "is_editable": {
                    "type": "boolean"
                },
                "is_public": {
                    "type": "boolean"
                },
                "label": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "app_controllers_admin_settings.UpdateSettingRequest": {
            "type": "object",
            "properties": {
                "value": {
                    "type": "string"
                }
            }
        },
...
        "app_controllers_user_profile.UpdateSettingsRequest": {
            "type": "object",
            "properties": {
                "language": {
                    "type": "string"
                },
                "notify_email": {
                    "type": "boolean"
                },
                "theme": {
                    "type": "string"
                }
            }
        },
...
        "fst_backend_app_models.SettingDTO": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "is_editable": {
                    "type": "boolean"
                },
                "is_public": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "value": {}
            }
        },
        "fst_backend_app_models.SettingsGroup": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fst_backend_app_models.SettingDTO"
                    }
                },
                "label": {
                    "type": "string"
                }
            }
        },
...
        "utils.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT认证令牌，格式: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}